Data.Text.Short.Unsafe

Description

Unsafe API

This module provides unsafe conversion functions

Documentation

fromShortByteStringUnsafe :: ShortByteString -> ShortText

\(\mathcal{O}(0)\) Construct ShortText from UTF-8 encoded ShortByteString

This operation has effectively no overhead, as it's currently merely a newtype-cast.

WARNING: Unlike the safe fromShortByteString conversion, this conversion is unsafe as it doesn't validate the well-formedness of the UTF-8 encoding.

fromByteStringUnsafe :: ByteString -> ShortText

\(\mathcal{O}(n)\) Construct ShortText from UTF-8 encoded ByteString

This operation is \(\mathcal{O}(n)\) because the ByteString needs to be copied into an unpinned ByteArray#.

WARNING: Unlike the safe fromByteString conversion, this conversion is unsafe as it doesn't validate the well-formedness of the UTF-8 encoding.